Add number

+++ Labs/src/Generic/Q3/TripleTestGenQ3.java
@@ -39,11 +39,11 @@ class Triple<T extends Number & Comparable<T>> {
 public class TripleTestGenQ3 {
 	public static void main(String[] args) {
 
-		Triple<Integer> t = new Triple<Integer>(8, 10, 7);
+		Triple<Double> t = new Triple<Double>(8.0, 10.0, 7.0);
 		System.out.println("Original Value1=" + t.readValue1());
 		System.out.println("Adding 3 Values=" + t.addValues());
 		System.out.println("Largest= " + t.largest());
-		t.setValue1(13);
+		t.setValue1(13.0);
 		System.out.println("Largest= " + t.largest());
 	}
 }
